Refactor: 수정 패키지 구조 수정.
- dto -> dto.board 로 변경.
- repository -> repository.board 로 변경.

Resolve: giantstep-study-board #15
