+++ src/main/java/com/giantstep/board/domain/board/dto/board/BoardSearchCondition.java
@@ -1,4 +1,4 @@
-package com.giantstep.board.domain.board.dto;
+package com.giantstep.board.domain.board.dto.board;
 
 import lombok.Getter;
 import lombok.Setter;
