+++ src/main/java/com/giantstep/board/domain/board/repository/board/BoardRepository.java
@@ -1,23 +1,17 @@
-package com.giantstep.board.domain.board.repository;
+package com.giantstep.board.domain.board.repository.board;
 
-import com.giantstep.board.domain.board.dto.BoardListDto;
-import com.giantstep.board.domain.board.dto.BoardOneDetailDto;
 import com.giantstep.board.domain.board.entity.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
-import org.springframework.data.repository.query.Param;
-
-import java.util.List;
 
 public interface BoardRepository extends JpaRepository<Board, Long>, BoardRepositoryCustom, QuerydslPredicateExecutor<Board> {
 
-//    @Query("select new com.giantstep.board.domain.board.dto.BoardListDto(" +
+//    @Query("select new com.giantstep.board.domain.board.dto.board.BoardListDto(" +
 //            "b.id, b.writer, b.title, b.updateDate)" +
 //            " from Board b order by b.updateDate desc")
 //    List<BoardListDto> findAllByBoardListDto();
 
-//    @Query("select new com.giantstep.board.domain.board.dto.BoardOneDetailDto(" +
+//    @Query("select new com.giantstep.board.domain.board.dto.board.BoardOneDetailDto(" +
 //            "b.id, b.writer, b.title, b.contents, b.updateDate)" +
 //            " from Board b where b.id = :boardId")
 //    BoardOneDetailDto findByBoardOneDetailDto(@Param("boardId") Long boardId);
