+++ src/test/java/com/giantstep/board/domain/board/constant/BoardStatusConverterTest.java
@@ -1,7 +1,7 @@
 package com.giantstep.board.domain.board.constant;
 
 import com.giantstep.board.domain.board.entity.Board;
-import com.giantstep.board.domain.board.repository.BoardRepository;
+import com.giantstep.board.domain.board.repository.board.BoardRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -14,7 +14,6 @@ import javax.persistence.Query;
 
 import static org.junit.jupiter.api.Assertions.*;
 /*
-
 @SpringBootTest
 @Transactional
 @Rollback(value = true)
@@ -25,11 +24,11 @@ class BoardStatusConverterTest {
 
     @PersistenceContext
     private EntityManager em;
+*/
 
-    */
-/** convert 테스트하기 위해서 리포지토리에 별도의 Querydsl 메소드를 만들지 않고 EntityManager 를 활용한 native query 를 활용해서 컨버터 검증 *//*
+/** convert 테스트하기 위해서 리포지토리에 별도의 Querydsl 메소드를 만들지 않고 EntityManager 를 활용한 native query 를 활용해서 컨버터 검증 */
 
-    @Test
+    /*@Test
     void attribute_컨버터_테스트() {
 
         //give
